+++ akka-persistence-testkit/src/test/scala/akka/persistence/testkit/scaladsl/ScalaDslUtils.scala
@@ -8,17 +8,21 @@ import akka.actor.typed.ActorRef
 import akka.persistence.testkit.{ Cmd, CommonUtils, EmptyState, Evt, Passivate, Recovered, Stopped, TestCommand }
 import akka.persistence.typed.PersistenceId
 import akka.persistence.typed.RecoveryCompleted
-import akka.persistence.typed.scaladsl.{ Effect, EventSourcedBehavior }
+import akka.persistence.typed.scaladsl.{ Effect, ReplicatedEventSourcedBehavior }
 
 trait ScalaDslUtils extends CommonUtils {
 
   def eventSourcedBehavior(pid: String, replyOnRecovery: Option[ActorRef[Any]] = None) =
-    EventSourcedBehavior[TestCommand, Evt, EmptyState](PersistenceId.ofUniqueId(pid), EmptyState(), (_, cmd) => {
-      cmd match {
-        case Cmd(data) => Effect.persist(Evt(data))
-        case Passivate => Effect.stop().thenRun(_ => replyOnRecovery.foreach(_ ! Stopped))
-      }
-    }, (_, _) => EmptyState()).snapshotWhen((_, _, _) => true).receiveSignal {
+    ReplicatedEventSourcedBehavior[TestCommand, Evt, EmptyState](
+      PersistenceId.ofUniqueId(pid),
+      EmptyState(),
+      (_, cmd) => {
+        cmd match {
+          case Cmd(data) => Effect.persist(Evt(data))
+          case Passivate => Effect.stop().thenRun(_ => replyOnRecovery.foreach(_ ! Stopped))
+        }
+      },
+      (_, _) => EmptyState()).snapshotWhen((_, _, _) => true).receiveSignal {
       case (_, RecoveryCompleted) => replyOnRecovery.foreach(_ ! Recovered)
     }
